+++ src/AdminGuesser.tsx
@@ -1,26 +1,27 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
 import {
   Admin,
   ComponentPropType,
-  Error as DefaultError,
+  I18nContextProvider,
   Loading,
-  TranslationProvider,
+  ThemeProvider,
   defaultI18nProvider,
 } from 'react-admin';
 import { createHashHistory, createMemoryHistory } from 'history';
-import { createTheme as /* tree-shaking no-side-effects-when-called */ createMuiTheme } from '@material-ui/core';
-import type { ComponentType } from 'react';
-import type { AdminProps, CustomRoutes, ErrorProps } from 'react-admin';
+import { ErrorBoundary } from 'react-error-boundary';
+import type { ComponentType, ErrorInfo } from 'react';
+import type { AdminProps, ErrorProps } from 'react-admin';
 import type { Resource } from '@api-platform/api-doc-parser';
 
-import ErrorBoundary from './ErrorBoundary';
 import IntrospectionContext from './IntrospectionContext';
 import ResourceGuesser from './ResourceGuesser';
 import SchemaAnalyzerContext from './SchemaAnalyzerContext';
-import { Layout } from './layout';
-import introspectReducer from './introspectReducer';
+import { Error as DefaultError, Layout, LoginPage, lightTheme } from './layout';
 import type { ApiPlatformAdminDataProvider, SchemaAnalyzer } from './types';
+import getRoutesAndResourcesFromNodes, {
+  getSingleChildFunction,
+} from './getRoutesAndResourcesFromNodes';
 
 export interface AdminGuesserProps extends AdminProps {
   dataProvider: ApiPlatformAdminDataProvider;
@@ -29,7 +30,7 @@ export interface AdminGuesserProps extends AdminProps {
 }
 
 interface AdminGuesserWithErrorProps extends AdminGuesserProps {
-  error: ComponentType<ErrorProps>;
+  error?: ComponentType<ErrorProps>;
 }
 
 interface AdminResourcesGuesserProps extends Omit<AdminProps, 'loading'> {
@@ -60,7 +61,6 @@ const displayOverrideCode = (resources: Resource[]) => {
  */
 export const AdminResourcesGuesser = ({
   // Admin props
-  customReducers = {},
   loadingPage: LoadingPage = Loading,
   admin: AdminEl = Admin,
   // Props
@@ -74,38 +74,32 @@ export const AdminResourcesGuesser = ({
     return <LoadingPage />;
   }
 
-  let resourceChildren = children;
-  if (!resourceChildren && resources) {
+  let adminChildren = children;
+  const { resources: resourceChildren, customRoutes } =
+    getRoutesAndResourcesFromNodes(children);
+  if (
+    !getSingleChildFunction(adminChildren) &&
+    resourceChildren.length === 0 &&
+    resources
+  ) {
     const guessResources = includeDeprecated
       ? resources
       : resources.filter((r) => !r.deprecated);
-    resourceChildren = guessResources.map((r) => (
-      <ResourceGuesser name={r.name} key={r.name} />
-    ));
+    adminChildren = [
+      ...customRoutes,
+      ...guessResources.map((r) => (
+        <ResourceGuesser name={r.name} key={r.name} />
+      )),
+    ];
     displayOverrideCode(guessResources);
   }
 
   return (
-    <AdminEl
-      customReducers={{ introspect: introspectReducer, ...customReducers }}
-      loading={LoadingPage}
-      {...rest}>
-      {resourceChildren}
+    <AdminEl loading={LoadingPage} {...rest}>
+      {adminChildren}
     </AdminEl>
   );
 };
-
-const defaultTheme = createMuiTheme({
-  palette: {
-    primary: {
-      contrastText: '#ffffff',
-      main: '#38a9b4',
-    },
-    secondary: {
-      main: '#288690',
-    },
-  },
-});
 
 const AdminGuesser = ({
   // Props for SchemaAnalyzerContext
@@ -115,10 +109,10 @@ const AdminGuesser = ({
   // Admin props
   dataProvider,
   history,
-  customRoutes = [],
   layout = Layout,
+  loginPage = LoginPage,
   loading: loadingPage,
-  theme = defaultTheme,
+  theme = lightTheme,
   // Other props
   children,
   ...rest
@@ -126,7 +120,6 @@ const AdminGuesser = ({
   const [resources, setResources] = useState<Resource[]>([]);
   const [loading, setLoading] = useState(true);
   const [, setError] = useState();
-  const [addedCustomRoutes, setAddedCustomRoutes] = useState<CustomRoutes>([]);
   const [introspect, setIntrospect] = useState(true);
   let adminHistory = history;
 
@@ -150,9 +143,8 @@ const AdminGuesser = ({
 
     dataProvider
       .introspect()
-      .then(({ data, customRoutes: introspectCustomRoutes = [] }) => {
+      .then(({ data }) => {
         setResources(data.resources ?? []);
-        setAddedCustomRoutes(introspectCustomRoutes);
         setIntrospect(false);
         setLoading(false);
       })
@@ -183,8 +175,8 @@ const AdminGuesser = ({
           loading={loading}
           dataProvider={dataProvider}
           history={adminHistory}
-          customRoutes={[...addedCustomRoutes, ...customRoutes]}
           layout={layout}
+          loginPage={loginPage}
           loadingPage={loadingPage}
           theme={theme}
           {...rest}>
@@ -202,37 +194,43 @@ AdminGuesser.propTypes = {
   authProvider: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   i18nProvider: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   history: PropTypes.object,
-  customReducers: PropTypes.object,
   customSagas: PropTypes.array,
   initialState: PropTypes.object,
   schemaAnalyzer: PropTypes.object.isRequired,
   children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
   theme: PropTypes.object,
   includeDeprecated: PropTypes.bool,
-  customRoutes: PropTypes.array,
 };
 /* eslint-enable tree-shaking/no-side-effects-in-initialization */
 
 const AdminGuesserWithError = ({
-  error,
+  error: ErrorComponent = DefaultError,
+  i18nProvider = defaultI18nProvider,
+  theme = lightTheme,
   ...props
 }: AdminGuesserWithErrorProps) => {
-  if (!props.i18nProvider) {
-    throw new Error('Missing i18nProvider');
-  }
+  const [errorInfo, setErrorInfo] = useState<ErrorInfo>();
+
+  const handleError = (_error: Error, info: ErrorInfo) => {
+    setErrorInfo(info);
+  };
+
+  const renderError = useCallback(
+    (fallbackRenderProps) => (
+      <ErrorComponent {...fallbackRenderProps} errorInfo={errorInfo} />
+    ),
+    [ErrorComponent, errorInfo],
+  );
 
   return (
-    <TranslationProvider i18nProvider={props.i18nProvider}>
-      <ErrorBoundary error={error}>
-        <AdminGuesser {...props} />
-      </ErrorBoundary>
-    </TranslationProvider>
+    <I18nContextProvider value={i18nProvider}>
+      <ThemeProvider theme={theme}>
+        <ErrorBoundary onError={handleError} fallbackRender={renderError}>
+          <AdminGuesser {...props} />
+        </ErrorBoundary>
+      </ThemeProvider>
+    </I18nContextProvider>
   );
-};
-
-AdminGuesserWithError.defaultProps = {
-  error: DefaultError,
-  i18nProvider: defaultI18nProvider,
 };
 
 AdminGuesserWithError.propTypes = {