+++ src/ResourceGuesser.tsx
@@ -1,27 +1,35 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Resource } from 'react-admin';
-import type { ResourceProps } from 'react-admin';
+import type { ResourceDefinition, ResourceProps } from 'react-admin';
 import ListGuesser from './ListGuesser';
 import CreateGuesser from './CreateGuesser';
 import EditGuesser from './EditGuesser';
 import ShowGuesser from './ShowGuesser';
 
 const ResourceGuesser = ({
-  list,
-  edit,
-  create,
-  show,
+  list = ListGuesser,
+  edit = EditGuesser,
+  create = CreateGuesser,
+  show = ShowGuesser,
   ...props
 }: ResourceProps) => (
-  <Resource
-    {...props}
-    create={undefined === create ? CreateGuesser : create}
-    edit={undefined === edit ? EditGuesser : edit}
-    list={undefined === list ? ListGuesser : list}
-    show={undefined === show ? ShowGuesser : show}
-  />
+  <Resource {...props} create={create} edit={edit} list={list} show={show} />
 );
+
+ResourceGuesser.raName = 'Resource';
+
+ResourceGuesser.registerResource = (
+  props: ResourceProps,
+): ResourceDefinition => ({
+  name: props.name,
+  icon: props.icon,
+  options: props.options,
+  hasList: true,
+  hasEdit: true,
+  hasCreate: true,
+  hasShow: true,
+});
 
 ResourceGuesser.propTypes = {
   name: PropTypes.string.isRequired,