+++ src/dataProvider/index.ts
@@ -1,0 +1,4 @@
+import useUpdateCache from './useUpdateCache';
+
+// eslint-disable-next-line import/prefer-default-export
+export { useUpdateCache };