+++ src/hydra/index.ts
@@ -1,6 +1,13 @@
 import dataProvider from './dataProvider';
 import fetchHydra from './fetchHydra';
 import HydraAdmin from './HydraAdmin';
+import type { HydraAdminProps } from './HydraAdmin';
 import schemaAnalyzer from './schemaAnalyzer';
 
-export { dataProvider, fetchHydra, HydraAdmin, schemaAnalyzer };
+export {
+  dataProvider,
+  fetchHydra,
+  HydraAdmin,
+  HydraAdminProps,
+  schemaAnalyzer,
+};